requirements.txt
streamlit
yfinance
pandas
numpy
plotly
scikit-learn
tensorflow-cpu==2.15.0
feedparser





